Sandoche/fse 509 jest config (#263)
