+++ apps/governance/package.json
@@ -1,6 +1,6 @@
 {
   "name": "governance-page",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "private": true,
   "scripts": {
     "dev": "next dev --port 3001",
@@ -10,7 +10,8 @@
     "lintunix": "next lint --format unix",
     "prepare": "husky install",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest --passWithNoTests",
-    "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --watch"
+    "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest --watch",
+    "test:coverage": "jest --coverage --passWithNoTests"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
@@ -53,9 +54,6 @@
   "devDependencies": {
     "@keplr-wallet/types": "^0.11.26",
     "@metamask/providers": "^10.2.1",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@types/jest": "^29.2.4",
     "@typescript-eslint/eslint-plugin": "^5.46.1",
     "@typescript-eslint/parser": "^5.46.1",
     "autoprefixer": "^10.4.13",
@@ -64,13 +62,11 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.31.11",
     "husky": "^8.0.2",
-    "jest": "^29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
     "lint-staged": "^13.1.0",
     "postcss": "^8.4.20",
     "prettier": "^2.8.1",
     "tailwindcss": "^3.2.4",
-    "ts-jest": "^29.1.0"
+    "jest-config-custom": "*"
   },
   "browserslist": [
     "last 3 chrome version",
