+++ package.json
@@ -15,6 +15,7 @@
     "clean": "turbo run clean",
     "test": "turbo run test",
     "test:watch": "turbo run test:watch",
+    "test:coverage": "turbo run test:coverage",
     "format": "prettier --write \"**/*.{ts,tsx,md}\""
   },
   "devDependencies": {
