+++ packages/helpers/package.json
@@ -1,6 +1,6 @@
 {
   "name": "helpers",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "sideEffects": [
     "**/*.css"
   ],
@@ -13,7 +13,10 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "check-types": "tsc --noEmit"
+    "check-types": "tsc --noEmit",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage"
   },
   "peerDependencies": {
     "react": "^18.2.0"
@@ -26,11 +29,9 @@
     "react": "^18.2.0",
     "tailwind-config": "workspace:*",
     "tsconfig": "workspace:*",
+    "jest-config-custom": "*",
     "tsup": "^6.1.3",
     "typescript": "^4.9.4",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@types/jest": "^29.2.4",
     "@typescript-eslint/eslint-plugin": "^5.46.1",
     "@typescript-eslint/parser": "^5.46.1",
     "autoprefixer": "^10.4.13",
@@ -38,8 +39,6 @@
     "eslint-plugin-jest": "^27.1.6",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.31.11",
-    "jest": "^29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
     "prettier": "^2.8.1"
   },
   "depedencies": {
