+++ packages/jest-config-custom/jest.packages.config.js
@@ -1,0 +1,8 @@
+const sharedConfig = require("./jest.shared.config.js");
+
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+    ...sharedConfig,
+    preset: "ts-jest",
+    testEnvironment: "node"
+}
