+++ packages/jest-config-custom/jest.shared.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    watchPlugins: [
+        'jest-watch-typeahead/filename',
+        'jest-watch-typeahead/testname',
+    ],
+}
