+++ packages/jest-config-custom/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "jest-config-custom",
+  "version": "1.0.0",
+  "private": true,
+  "devDependencies": {
+    "jest": "^29.5.0",
+    "ts-jest": "^29.1.0",
+    "@types/jest": "^29.5.0",
+    "jest-watch-typeahead": "^2.2.2",
+    "jest-environment-jsdom": "^29.5.0"
+  }
+}
