+++ packages/ui-helpers/jest.config.js
@@ -1,0 +1,5 @@
+const sharedConfig = require('jest-config-custom/jest.packages.config');
+
+module.exports = {
+    ...sharedConfig
+}
