+++ turbo.json
@@ -12,6 +12,7 @@
     "eslint-config-custom#build": {},
     "tailwind-config#build": {},
     "tsconfig#build": {},
+    "jest-config-evmos#build": {},
     "constants-helper#build": {},
     "icons#build": {},
     "helpers#build": {},
@@ -48,6 +49,9 @@
     },
     "test": {},
     "test:watch": {
+      "cache": false
+    },
+    "test:coverage": {
       "cache": false
     }
   }
